requirements.txt
aiohttp
aiosignal
async-timeout
attrs
cachetools
certifi
chardet
charset-normalizer
click
colorama
discord.py==2.2.2
ffxivweather
frozenlist
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
Google-Images-Search
googleapis-common-protos
httplib2
idna
multidict
protobuf
pyasn1
pyasn1-modules
pyfiglet
pyparsing
python-resize-image
pytz
requests
rsa
six
termcolor
udpy
uritemplate
urllib3
yarl
